+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/notificationBox/primaryNotification/recyclerview/viewHolder/PrimaryNotificationViewHolder.kt
@@ -1,0 +1,8 @@
+package com.emmsale.presentation.ui.notificationBox.primaryNotification.recyclerview.viewHolder
+
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+
+sealed class PrimaryNotificationViewHolder(
+    parent: ViewGroup,
+) : RecyclerView.ViewHolder(parent)
