--- android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/notificationBox/primaryNotification/recyclerview/viewholder/RecentNotificationViewHolder.kt
@@ -1,9 +1,0 @@
-package com.emmsale.presentation.ui.notificationBox.primaryNotification.recyclerview.viewholder
-
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-import com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate.PrimaryNotificationUiState
-
-abstract class RecentNotificationViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-    abstract fun bind(item: PrimaryNotificationUiState)
-}
