--- android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/notificationBox/primaryNotification/uistate/PrimaryNotificationsUiState.kt
@@ -1,16 +1,0 @@
-package com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate
-
-data class PrimaryNotificationsUiState(
-    val notifications: List<PrimaryNotificationUiState> = emptyList(),
-    val isLoading: Boolean = false,
-    val isError: Boolean = false,
-) {
-    fun deleteNotificationById(notificationId: Long): PrimaryNotificationsUiState =
-        copy(notifications = notifications.filterNot { it.id == notificationId })
-
-    val notificationIds = notifications.map(PrimaryNotificationUiState::id)
-
-    companion object {
-        val EMPTY: PrimaryNotificationsUiState = PrimaryNotificationsUiState()
-    }
-}
