--- android/2023-emmsale/app/src/main/res/layout/item_past_notification_header.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-        <variable
-            name="onDeleteAllClick"
-            type="kotlin.jvm.functions.Function0" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@drawable/bg_primarynotification_header"
-        android:paddingHorizontal="17dp"
-        android:paddingTop="30dp"
-        android:paddingBottom="20dp">
-
-        <TextView
-            android:id="@+id/tv_past_notification_title"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:text="@string/primarynotification_past_notification_header_title"
-            android:textColor="@color/black"
-            android:textSize="14sp"
-            android:textStyle="bold"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/btn_delete_all"
-            android:layout_width="wrap_content"
-            android:layout_height="0dp"
-            android:background="?attr/selectableItemBackgroundBorderless"
-            android:clickable="true"
-            android:focusable="true"
-            android:gravity="center"
-            android:onClick="@{() -> onDeleteAllClick.invoke()}"
-            android:text="@string/primarynotification_past_notification_delete_all"
-            android:textAppearance="@style/KerdyTextStyleGrayBold12"
-            android:textColor="@color/light_gray"
-            app:layout_constraintBottom_toBottomOf="@+id/tv_past_notification_title"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@+id/tv_past_notification_title" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
