+++ android/2023-emmsale/app/src/main/res/layout/item_primarynotification_child_comment_notification.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="notification"
-            type="com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate.ChildCommentNotificationUiState1" />
+            type="com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate.ChildCommentNotificationUiState" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
