+++ android/2023-emmsale/app/src/main/res/layout/item_primarynotification_interest_event_notification.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="notification"
-            type="com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate.InterestEventNotificationUiState1" />
+            type="com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate.InterestEventNotificationUiState" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
